requirements.txt
Flask==2.2.1  
Flask-SQLAlchemy==3.0.3
Werkzeug==2.2.2
authlib==0.15.4
pymysql
requests